Migrate billing methods to autoloaded file
