+++ includes/common.php
@@ -582,66 +582,6 @@ function formatStorage($value, $round = '2', $sf = '3')
     return $value;
 }
 
-function format_si($value, $round = '2', $sf = '3')
-{
-    $neg = 0;
-    if ($value < "0") {
-        $neg = 1;
-        $value = $value * -1;
-    }
-
-    if ($value >= "0.1") {
-        $sizes = array('', 'k', 'M', 'G', 'T', 'P', 'E');
-        $ext = $sizes[0];
-        for ($i = 1; (($i < count($sizes)) && ($value >= 1000)); $i++) {
-            $value = $value / 1000;
-            $ext  = $sizes[$i];
-        }
-    } else {
-        $sizes = array('', 'm', 'u', 'n', 'p');
-        $ext = $sizes[0];
-        for ($i = 1; (($i < count($sizes)) && ($value != 0) && ($value <= 0.1)); $i++) {
-            $value = $value * 1000;
-            $ext  = $sizes[$i];
-        }
-    }
-
-    if ($neg == 1) {
-        $value = $value * -1;
-    }
-
-        return number_format(round($value, $round), $sf, '.', '').$ext;
-}
-
-function format_bi($value, $round = '2', $sf = '3')
-{
-    if ($value < "0") {
-        $neg = 1;
-        $value = $value * -1;
-    }
-    $sizes = array('', 'k', 'M', 'G', 'T', 'P', 'E');
-    $ext = $sizes[0];
-    for ($i = 1; (($i < count($sizes)) && ($value >= 1024)); $i++) {
-        $value = $value / 1024;
-        $ext  = $sizes[$i];
-    }
-
-    if ($neg) {
-        $value = $value * -1;
-    }
-
-    return number_format(round($value, $round), $sf, '.', '').$ext;
-}
-
-function format_number($value, $base = '1000', $round = 2, $sf = 3)
-{
-    if ($base == '1000') {
-        return format_si($value, $round, $sf);
-    } else {
-        return format_bi($value, $round, $sf);
-    }
-}
-
 function is_valid_hostname($hostname)
 {
     // The Internet standards (Request for Comments) for protocols mandate that
